backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
requests==2.31.0
pyjwt==2.8.0
structlog==23.2.0
python-multipart==0.0.6
cryptography==41.0.8

# Optional dependencies
openai==1.3.8  # For LLM integration

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1